Add support for closed testsolving
If a puzzle is tagged as closed testsolving, then it doesn't show up in
the list of testsolveable puzzles for unprivileged users. This means
that a testsolve coordinator or EIC has to explicitly create testsolve
sessions for these puzzles.

Additionally, to enable that, make it possible to start testsolve
sessions with a list of participants, which (if you're a testsolve
coordinator) doesn't have to include yourself. This allows an
improvement to the testsolver finder UI such that it can actually create
testsolve sessions with everyone, instead of just  yourself.