Add support for "late" testsolving of puzzles
If a puzzle has passed testsolving, is not a meta, and isn't marked for
closed testsolving, then we open it up for "late" testsolving. These
testsolves are meant to be fun for people solving them rather than
useful writing feedback, and we message that pretty strongly. We also
don't count late testsolves in users' stats.